microservicio_reportes/requirements.txt
Flask==2.3.2
celery==5.3.1
redis==4.5.5
# WeasyPrint requires system packages on Windows (pango/cairo). Keep optional.
# weasyprint==57.0
pandas==2.2.3
openpyxl==3.1.2
python-dotenv==1.0.0
boto3==1.28.55
reportlab==4.0.0

# Optional for local dev
# gunicorn
requests==2.31.0
